Add files via upload
The file 'replication.ipynb' contains a detailed Jupyter Notebook that outlines the procedural steps and analytical methods used to produce the results presented in Section 4 of the paper titled 'Machine Learning and Feature Selection: Applications in Economics and Climate Change'. This notebook includes code, commentary, and visualizations that guide the reader through the replication of my findings.